Use block number instead of u64 for import counter

+++ client/consensus/common/src/tests.rs
@@ -583,7 +583,7 @@ fn restore_limit_monitor() {
 		hashes
 			.iter()
 			.filter(|hash| **hash != block13.header.hash())
-			.all(|hash| *number as u64 == *monitor.freshness.get(hash).unwrap())
+			.all(|hash| *number == *monitor.freshness.get(hash).unwrap())
 	}));
 	assert_eq!(*monitor.freshness.get(&block13.header.hash()).unwrap(), monitor.import_counter - 1);
 }
